+++ source/Internal/Framework/Module/TemplateExtension/TemplateBlockNotFoundException.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\EshopCommunity\Internal\Framework\Module\TemplateExtension;
+
+class TemplateBlockNotFoundException extends \Exception
+{
+
+}